#98 - Create samples for BatchService testing - separating batch service files - added api-jee-batch module - added missing dependencies after coffee 2.6.0 BOM separation - jpamodelgen 6.1.6.Final -> 6.2.13.Final
