+++ etc/config/META-INF/microprofile-config-jdbc.yml
@@ -1,0 +1,5 @@
+coffee:
+  config:
+    datasource:
+      default:
+        name: icellmobilsoftDS
