+++ model/pom.xml
@@ -10,7 +10,7 @@
     <artifactId>model</artifactId>
 
     <properties>
-        <hibernate.jpamodelgen.version>6.1.6.Final</hibernate.jpamodelgen.version>
+        <hibernate.jpamodelgen.version>6.2.13.Final</hibernate.jpamodelgen.version>
     </properties>
 
     <dependencyManagement>
