+++ sample/sample-jpa-batch-service/pom.xml
@@ -20,10 +20,15 @@
             <artifactId>common-system-jpa</artifactId>
             <exclusions>
                 <exclusion>
-                    <groupId>h2.database</groupId>
-                    <artifactId>h2</artifactId>
+                    <groupId>hu.icellmobilsoft.sampler.api</groupId>
+                    <artifactId>api-jee</artifactId>
                 </exclusion>
             </exclusions>
+        </dependency>
+
+        <dependency>
+            <groupId>hu.icellmobilsoft.sampler.api</groupId>
+            <artifactId>api-jee-batch</artifactId>
         </dependency>
 
         <!-- oracle -->
@@ -58,7 +63,6 @@
             <resource>
                 <directory>../../etc/config</directory>
                 <includes>
-                    <include>META-INF/persistence.xml</include>
                     <include>META-INF/microprofile-config*</include>
                 </includes>
                 <filtering>true</filtering>
