+++ sample/sample-jpa-batch-service/src/main/java/hu/icellmobilsoft/sampler/sample/jpa/batch/service/rest/EmptyEntityRest.java
@@ -24,7 +24,7 @@ import jakarta.inject.Inject;
 
 import hu.icellmobilsoft.coffee.dto.common.commonservice.BaseRequest;
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
-import hu.icellmobilsoft.sampler.api.jee.rest.batch.IEmptyEntityRest;
+import hu.icellmobilsoft.sampler.api.jee.batch.rest.IEmptyEntityRest;
 import hu.icellmobilsoft.sampler.common.system.rest.rest.BaseRestService;
 import hu.icellmobilsoft.sampler.dto.sample.batch.emptyentity.EmptyEntityResponse;
 import hu.icellmobilsoft.sampler.sample.jpa.batch.service.action.EmptyEntityAction;
