+++ testsuite/ts-sample/ts-sample-jpa-batch-service/ts-sample-jpa-batch-service-java-base-types-rest/src/test/java/hu/icellmobilsoft/ts/sample/jpa/batch/service/java/base/types/rest/DeleteAllJavaBaseTypesEntitiesWithBatchServiceIT.java
@@ -31,7 +31,7 @@ import hu.icellmobilsoft.coffee.dto.common.commonservice.BaseResponse;
 import hu.icellmobilsoft.coffee.dto.common.commonservice.FunctionCodeType;
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
 import hu.icellmobilsoft.roaster.api.TestSuiteGroup;
-import hu.icellmobilsoft.sampler.api.jee.rest.batch.IJavaBaseTypesRest;
+import hu.icellmobilsoft.sampler.api.jee.batch.rest.IJavaBaseTypesRest;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javabasetypes.JavaBaseTypesInsertType;
 import hu.icellmobilsoft.ts.sample.jpa.batch.service.java.base.types.rest.base.BaseJavaBaseTypesRestIT;
