+++ testsuite/ts-sample/ts-sample-jpa-batch-service/ts-sample-jpa-batch-service-java-data-rest/src/test/java/hu/icellmobilsoft/ts/sample/jpa/batch/service/java/data/rest/PostInsertJavaDataEntityWithBatchServiceIT.java
@@ -30,7 +30,7 @@ import org.junit.jupiter.api.TestMethodOrder;
 import hu.icellmobilsoft.coffee.dto.common.commonservice.FunctionCodeType;
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
 import hu.icellmobilsoft.roaster.api.TestSuiteGroup;
-import hu.icellmobilsoft.sampler.api.jee.rest.batch.IJavaDataRest;
+import hu.icellmobilsoft.sampler.api.jee.batch.rest.IJavaDataRest;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javadata.JavaDataResponse;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javadata.JavaDataType;
 import hu.icellmobilsoft.ts.sample.jpa.batch.service.java.data.rest.base.BaseJavaDataRestIT;
