+++ testsuite/ts-sample/ts-sample-jpa-batch-service/ts-sample-jpa-batch-service-java-date-and-time-rest/src/test/java/hu/icellmobilsoft/ts/sample/jpa/batch/service/java/date/and/time/rest/base/BaseJavaDateAndTimeRestIT.java
@@ -22,7 +22,7 @@ package hu.icellmobilsoft.ts.sample.jpa.batch.service.java.date.and.time.rest.ba
 import java.time.OffsetDateTime;
 
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
-import hu.icellmobilsoft.sampler.api.jee.rest.batch.IJavaDateAndTimeRest;
+import hu.icellmobilsoft.sampler.api.jee.batch.rest.IJavaDateAndTimeRest;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javadata.JavaDataResponse;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javadateandtime.JavaDateAndTimeInsertRequest;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javadateandtime.JavaDateAndTimeInsertType;
