+++ testsuite/ts-sample/ts-sample-jpa-batch-service/ts-sample-jpa-batch-service-java-enum-types-rest/src/test/java/hu/icellmobilsoft/ts/sample/jpa/batch/service/java/enumtypes/rest/PutUpdateJavaEnumTypesEntityWithBatchServiceIT.java
@@ -31,7 +31,7 @@ import hu.icellmobilsoft.coffee.dto.common.commonservice.FunctionCodeType;
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
 import hu.icellmobilsoft.coffee.tool.utils.enums.EnumUtil;
 import hu.icellmobilsoft.roaster.api.TestSuiteGroup;
-import hu.icellmobilsoft.sampler.api.jee.rest.batch.IJavaEnumTypesRest;
+import hu.icellmobilsoft.sampler.api.jee.batch.rest.IJavaEnumTypesRest;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javaenumtypes.JavaEnumTypesInsertType;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javaenumtypes.JavaEnumTypesResponse;
 import hu.icellmobilsoft.sampler.dto.sample.batch.javaenumtypes.JavaEnumTypesType;
