+++ testsuite/ts-sample/ts-sample-jpa-batch-service/ts-sample-jpa-batch-service-jpa-converter-entity-rest/src/test/java/hu/icellmobilsoft/ts/sample/jpa/batch/service/jpa/converter/entity/rest/DeleteAllJpaConverterEntityEntitiesWithBatchServiceIT.java
@@ -34,7 +34,7 @@ import hu.icellmobilsoft.coffee.dto.common.commonservice.BaseResponse;
 import hu.icellmobilsoft.coffee.dto.common.commonservice.FunctionCodeType;
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
 import hu.icellmobilsoft.roaster.api.TestSuiteGroup;
-import hu.icellmobilsoft.sampler.api.jee.rest.batch.IJpaConverterEntityRest;
+import hu.icellmobilsoft.sampler.api.jee.batch.rest.IJpaConverterEntityRest;
 import hu.icellmobilsoft.ts.sample.jpa.batch.service.jpa.converter.entity.rest.base.BaseJpaConverterEntityRestIT;
 
 /**
