+++ testsuite/ts-sample/ts-sample-jpa-batch-service/ts-sample-jpa-batch-service-jpa-converter-entity-rest/src/test/java/hu/icellmobilsoft/ts/sample/jpa/batch/service/jpa/converter/entity/rest/PutUpdateJpaConverterEntityEntityWithBatchServiceIT.java
@@ -33,7 +33,7 @@ import org.junit.jupiter.api.TestMethodOrder;
 import hu.icellmobilsoft.coffee.dto.common.commonservice.FunctionCodeType;
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
 import hu.icellmobilsoft.roaster.api.TestSuiteGroup;
-import hu.icellmobilsoft.sampler.api.jee.rest.batch.IJpaConverterEntityRest;
+import hu.icellmobilsoft.sampler.api.jee.batch.rest.IJpaConverterEntityRest;
 import hu.icellmobilsoft.sampler.dto.sample.batch.jpaconverterentity.JpaConverterEntityResponse;
 import hu.icellmobilsoft.sampler.dto.sample.batch.jpaconverterentity.JpaConverterEntityType;
 import hu.icellmobilsoft.sampler.dto.sample.batch.jpaconverterentity.JpaConverterEntityUpdateRequest;
