+++ testsuite/ts-sample/ts-sample-jpa-batch-service/ts-sample-jpa-batch-service-jpa-converter-entity-rest/src/test/java/hu/icellmobilsoft/ts/sample/jpa/batch/service/jpa/converter/entity/rest/base/BaseJpaConverterEntityRestIT.java
@@ -22,7 +22,7 @@ package hu.icellmobilsoft.ts.sample.jpa.batch.service.jpa.converter.entity.rest.
 import java.time.OffsetDateTime;
 
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
-import hu.icellmobilsoft.sampler.api.jee.rest.batch.IJpaConverterEntityRest;
+import hu.icellmobilsoft.sampler.api.jee.batch.rest.IJpaConverterEntityRest;
 import hu.icellmobilsoft.sampler.dto.sample.batch.jpaconverterentity.JpaConverterEntityInsertRequest;
 import hu.icellmobilsoft.sampler.dto.sample.batch.jpaconverterentity.JpaConverterEntityInsertType;
 import hu.icellmobilsoft.sampler.dto.sample.batch.jpaconverterentity.JpaConverterEntityResponse;
